pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "openpdf"
version = "0.1.0"
description = "A commercially-free drop-in replacement for PyMuPDF (fitz)"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10,<3.12"
authors = [{ name = "OpenPDF Contributors" }]
keywords = ["pdf", "pymupdf", "fitz", "text-extraction", "pdf-rendering"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Multimedia :: Graphics",
]
dependencies = [
    "pypdfium2>=4.25,<5",
    "pikepdf>=8,<10",
    "pdfminer.six>=20221105",
    "Pillow>=10,<11",
    "reportlab>=4,<5",
    "fonttools>=4.40",
]

[project.optional-dependencies]
ocr = ["pytesseract>=0.3.10"]
dev = [
    "pytest>=7",
    "pytest-cov>=4",
    "ruff>=0.1",
    "mypy>=1.5",
]

[project.urls]
Repository = "https://github.com/letstrythisshit/openpdf"

[tool.setuptools.packages.find]
where = ["."]
include = ["openpdf*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = false
ignore_missing_imports = true
